requirements.txt
python3-flask
python3-itsdangerous
python3-pyinotify
python3-werkzeug
